아이템 이펙트 스크립트 분리
