+++ Assets/Scripts/Item/ItemEffect/ItemEffectStat.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b8814312271d17940b0d8d025e629241
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
